requirements.txt.txt
flask
gunicorn
botbuilder-core
botbuilder-schema
aiohttp